delete parser changes

+++ Combined/SQL_Parser/src/Screens/Delete.java
@@ -6,6 +6,8 @@ import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
+import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import javax.swing.JOptionPane;
@@ -17,11 +19,10 @@ import org.json.simple.parser.ParseException;
 public class Delete {	
 
 	String tableName;
-	List<String> whereConditions;
+	List<WhereClause> whereConditions;
 	boolean conditionFlag = false;
 	HashMap<String,String> conditions = new HashMap<String,String>();
 	String conditionOp;
-	//List<WhereClause> conditions = new ArrayList<WhereClause>();
 
 
 	public boolean parse(String sql) {
@@ -38,12 +39,6 @@ public class Delete {
 		// where clause exists
 
 		int whereIndex = (sql.indexOf("WHERE") != -1) ? sql.indexOf("WHERE") : sql.indexOf("where");
-
-		/*if(whereIndex == -1){			
-			JOptionPane.showMessageDialog(null, "Invalid Syntax,Missing WHERE clause", "Error", JOptionPane.ERROR_MESSAGE);
-			return;	
-		}*/
-
 
 		// check if from is at position 2
 		String [] tokens = sql.split("\\s+");
@@ -103,11 +98,12 @@ public class Delete {
 
 			if("".equals(condition)){
 
-				JOptionPane.showMessageDialog(null, "Invalid Syntax, Missing conditions", "Error", JOptionPane.ERROR_MESSAGE);
+				JOptionPane.showMessageDialog(null, "Invalid Syntax, Missing conditions after WHERE clause", "Error", JOptionPane.ERROR_MESSAGE);
 				return false;	
 
 			}else if(condition.startsWith("and") || condition.startsWith("or") || condition.endsWith("and") || condition.endsWith("or")){		    	  
-				JOptionPane.showMessageDialog(null, "Invalid Syntax, Missing conditions after And or or", "Error", JOptionPane.ERROR_MESSAGE);
+				
+				JOptionPane.showMessageDialog(null, "Invalid Syntax, Missing conditions before/after AND or OR", "Error", JOptionPane.ERROR_MESSAGE);
 				return false;	
 
 			}else{
@@ -124,19 +120,47 @@ public class Delete {
 
 
 		//if syntax is correct, check if tableName and columnName exists
-		/*if(!GlobalUtil.validateTableName(this.tableName)){
+		if(!GlobalUtil.validateTableName(this.tableName)){
 			JOptionPane.showMessageDialog(null, "Invalid Syntax: No such table exists", "Error", JOptionPane.ERROR_MESSAGE);
 			return false;
 
-		}*/
-
+		}
+	
+		/* 
+		 * validate if colName exists or not
+		 */
+		boolean validColNames = validateColNames();
+		
+		if(!validColNames){
+			JOptionPane.showMessageDialog(null, "Invalid Column Name", "Error", JOptionPane.ERROR_MESSAGE);
+			return false;			
+		}
+		
 		///if everything is good, delete records	
 		deleteRecords();
 		return true;
 
 	}
 
+	
+	
+	public boolean validateColNames(){
+		
+	    List<String> colNames =  new ArrayList<String>();
+		
+		for(WhereClause clause : this.whereConditions){			
+			  colNames.add(clause.attribute1);
+		}
+		
+		if(!GlobalUtil.validateColumnNames(colNames, this.tableName))
+			 return false;
+		
+		return true;
+		
+	}
+	
 
+	
 	public boolean fetchWhereClause(String sql){
 
 		if(sql.indexOf("=",0) == -1 && sql.indexOf("<",0) == -1 && sql.indexOf(">",0) == -1){			 
@@ -144,8 +168,7 @@ public class Delete {
 		}
 		
 		System.out.println("inside fetchWhereClause");
-		
-		
+				
 		String [] firstConditionArr = sql.split("\\s+");
 		
 		boolean conditionFlag = true;
@@ -188,13 +211,15 @@ public class Delete {
 				  
 				  System.out.println("colName:"+colName);
 				  System.out.println("colVal:"+colVal);
+				 
 				  
-				  // validate that colVal is inside quotes
+				  //conditions.put(colName.trim(), colVal.trim());
+				  WhereClause where = new WhereClause();
+				  where.attribute1 = colName;
+				  where.attribute2 = colVal;
+				  where.operation = operator;
 				  
-				 // if(!validateColVal(colVal))
-					//   return false;
-				  
-				  conditions.put(colName.trim(), colVal.trim());
+				  this.whereConditions.add(where);
 				 				  
 				  conditionFlag = false;
 				  andOrFlag = true;
@@ -247,16 +272,68 @@ public class Delete {
 
 				JSONObject temp = (JSONObject) parser.parse(headers.get(i).toString());
 
-				// iterate through each condition and check
-
-				/*for(Map.Entry<String, String> entryMap : conditions.entrySet()){
-
-					String dataVal = (String)temp.get(entryMap.getKey());
-					if(dataVal.equals(entryMap.getValue())){
-						headers.remove(i);
-					}					 					
-				}*/
-
+				//iterate through each condition and check
+				boolean allConditionsMatch = true;
+				
+				for(WhereClause whereClause: this.whereConditions){
+					
+					 String colName = whereClause.attribute1;
+					 String colVal = whereClause.attribute2;
+					 
+					 String value = (String)temp.get(colName);
+					 
+					 char operator = whereClause.operation;
+					 
+					 if(operator == '>'){
+						  System.out.println("Inside > ");                         						 
+						  BigDecimal searchVal = new BigDecimal(colVal);
+		                  BigDecimal actualVal = new BigDecimal(value);
+		                  
+		                  if(actualVal.compareTo(searchVal) <= 0){
+		                	  allConditionsMatch = false;
+	                		  break;
+	                	 }	
+		                	 	
+					 }else if(operator == '<'){
+						  System.out.println("Inside < ");
+                        						 
+						  BigDecimal searchVal = new BigDecimal(colVal);
+		                  BigDecimal actualVal = new BigDecimal(value);
+		                  
+		                  if(actualVal.compareTo(searchVal) <= 0){
+		                	  allConditionsMatch = false;
+	                		  break;
+	                	 }
+					 }else if(operator == '=') {
+						
+						 System.out.println("Inside ="); 
+						 
+						 BigDecimal searchVal = new BigDecimal(colVal);
+	                	 BigDecimal actualVal = new BigDecimal(value); 
+	                	 
+						 if(searchVal.compareTo(actualVal) != 0){
+							 allConditionsMatch = false;
+							 break;								
+						 }		
+					 }else{
+						 
+						 if(colVal.equalsIgnoreCase(value)){						 
+							 continue;
+						 }else{
+							 allConditionsMatch = false;
+							 break;
+						 }
+												 
+					 }				
+				}
+				
+				if(allConditionsMatch){
+					
+					// delete the record from the json file
+				
+					
+				}
+			
 			}
 
 			// write the file back to disk
